Refactor ACS command group
[noissue]
