+++ pulpcore/cli/file/__init__.py
@@ -2,9 +2,10 @@ from typing import Any
 
 import click
 from pulp_glue.common.i18n import get_translation
+from pulp_glue.file.context import PulpFileACSContext, PulpFileRemoteContext
 
+from pulpcore.cli.common.acs import acs_command
 from pulpcore.cli.common.generic import pulp_group
-from pulpcore.cli.file.acs import acs
 from pulpcore.cli.file.content import content
 from pulpcore.cli.file.distribution import distribution
 from pulpcore.cli.file.publication import publication
@@ -26,5 +27,10 @@ def mount(main: click.Group, **kwargs: Any) -> None:
     file_group.add_command(publication)
     file_group.add_command(distribution)
     file_group.add_command(content)
-    file_group.add_command(acs)
+    file_group.add_command(
+        acs_command(
+            acs_contexts={"file": PulpFileACSContext},
+            remote_context_table={"file:file": PulpFileRemoteContext},
+        )
+    )
     main.add_command(file_group)
