+++ pulpcore/cli/rpm/__init__.py
@@ -2,9 +2,10 @@ from typing import Any
 
 import click
 from pulp_glue.common.i18n import get_translation
+from pulp_glue.rpm.context import PulpRpmACSContext, PulpRpmRemoteContext, PulpUlnRemoteContext
 
+from pulpcore.cli.common.acs import acs_command
 from pulpcore.cli.common.generic import pulp_group
-from pulpcore.cli.rpm.acs import acs
 from pulpcore.cli.rpm.comps import comps_upload
 from pulpcore.cli.rpm.content import content
 from pulpcore.cli.rpm.distribution import distribution
@@ -27,6 +28,11 @@ def mount(main: click.Group, **kwargs: Any) -> None:
     rpm.add_command(publication)
     rpm.add_command(distribution)
     rpm.add_command(content)
-    rpm.add_command(acs)
+    rpm.add_command(
+        acs_command(
+            acs_contexts={"rpm": PulpRpmACSContext},
+            remote_context_table={"rpm:rpm": PulpRpmRemoteContext, "rpm:uln": PulpUlnRemoteContext},
+        )
+    )
     rpm.add_command(comps_upload)
     main.add_command(rpm)
